Remove calculate_triangle_face_normal and face_tools.rs
